Map app, current location function added.
