+++ src/app/map-page/page.tsx
@@ -1,13 +1,14 @@
 import dynamic from "next/dynamic";
 import React from "react";
 
-import EventsExample from "../../components/events"
+import Map from "../../components/map/map"
+// import { currentPos } from "../../components/map/currentPosition"
 
 // Mapを呼び出す
 export default function MapPage() {
     const Map = React.useMemo(
         () =>
-        dynamic(() => import("../../components/map"), {
+        dynamic(() => import("../../components/map/map"), {
             loading: () => <p>A map is loading</p>,
             ssr: false,
         }),
@@ -15,10 +16,9 @@ export default function MapPage() {
     );
     return (
         <div>
+            {/* <p><button onClick={currentPos}>現在地を取得する</button></p> */}
             <Map />
             
-            EventsExample
-            <EventsExample />
         </div>
     )
 }
